fixing issue w/ app/env names w/ spaces. Adding lock file to avoid starting multiple agents.

+++ startup/stackify.sh
@@ -9,6 +9,7 @@ readonly STACKIFY_JAVA_OPTS="-XX:+UseSerialGC -Xmx192m"
 readonly STACKIFY_JAVA_JAR="${STACKIFY_AGENT_INSTALL_PATH}/stackify-agent.jar"
 readonly STACKIFY_JAVA_MAIN_CLASS="com.stackify.agent.AgentMain";
 readonly STACKIFY_AGENT_LOG="${STACKIFY_AGENT_INSTALL_PATH}/log/stackify-agent.log"
+readonly STACKIFY_HEROKU_LOCK_FILE="${STACKIFY_AGENT_INSTALL_PATH}/stackify-heroku.lock"
 
 # update stackify-agent configuration
 sed -i 's:^[ \t]*sudoDisabled[ \t]*=\([ \t]*.*\)$:sudoDisabled=true:' ${CONFIG_FILE}
@@ -21,12 +22,12 @@ fi
 
 if [ ! -z "$STACKIFY_APPLICATION_NAME" ]; then
     echo "STACKIFY_APPLICATION_NAME: ${STACKIFY_APPLICATION_NAME}"
-    sed -i 's:^[ \t]*deviceAlias[ \t]*=\([ \t]*.*\)$:deviceAlias='\"${STACKIFY_APPLICATION_NAME}\"':' ${CONFIG_FILE}
+    sed -i 's:^[ \t]*deviceAlias[ \t]*=\([ \t]*.*\)$:deviceAlias='\""${STACKIFY_APPLICATION_NAME}"\"':' ${CONFIG_FILE}
 fi
 
 if [ ! -z "$STACKIFY_ENVIRONMENT_NAME" ]; then
     echo "STACKIFY_ENV: ${STACKIFY_ENVIRONMENT_NAME}"
-    sed -i 's:^[ \t]*environment[ \t]*=\([ \t]*.*\)$:environment='\"${STACKIFY_ENVIRONMENT_NAME}\"':' ${CONFIG_FILE}
+    sed -i 's:^[ \t]*environment[ \t]*=\([ \t]*.*\)$:environment='\""${STACKIFY_ENVIRONMENT_NAME}"\"':' ${CONFIG_FILE}
 fi
 
 # set all profilers to use HTTP transport
@@ -35,7 +36,12 @@ export STACKIFY_TRANSPORT="agent_http"
 # start Stackify Linux Agent in background
 export STACKIFY_ROOT_FOLDER="${STACKIFY_HOME}"
 cd $STACKIFY_AGENT_INSTALL_PATH
-nohup $STACKIFY_JAVA_EXEC $STACKIFY_JAVA_OPTS -cp $STACKIFY_JAVA_JAR $STACKIFY_JAVA_MAIN_CLASS &
+
+# start if no lock file is present
+if [ ! -f "$STACKIFY_HEROKU_LOCK_FILE" ]; then
+    nohup $STACKIFY_JAVA_EXEC $STACKIFY_JAVA_OPTS -cp $STACKIFY_JAVA_JAR $STACKIFY_JAVA_MAIN_CLASS &
+    touch $STACKIFY_HEROKU_LOCK_FILE
+fi
 
 # set back to home directory
 cd $HOME
